requirements.txt
# Python dependencies
# Core dependencies
transformers==4.38.2
tokenizers>=0.14,<0.19
accelerate>=0.33.0
sentencepiece>=0.2.0
protobuf>=3.20.0
torch>=2.1.0
torchvision>=0.16.0
safetensors>=0.4.1

# DeepSeek-VL2 support
timm>=0.9.16
attrdict>=2.0.1
einops>=0.7.0

# Vision and image processing
pillow>=10.0.0
pdf2image>=1.16.0
pymupdf>=1.23.0

# Math and validation
sympy>=1.12
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
pyyaml>=6.0
requests>=2.31.0
imageio-ffmpeg>=0.5.1

# API hosting
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
streamlit>=1.39.0
